// src/enum/courriercategory.enum.ts
/**
 * Project gestion-courrier-api
 * File courriertype.enum
 * Path src/enum
 * Created by BRICE ZELE
 * Date: 08/03/2022
 */
export enum CourrierCategory {
  ARRETE = "arrete",
  COMPTE_RENDU = "compte_rendu",
  COURRIER = "courrier",
  DEPOT_PLAINTE = "depot_plainte",
  FACTURE = "facture",
  FAX = "fax",
  FICHE_SIGNALEMENT = "fiche_signalement",
  INTERVENTION = "intervention",
  MAIL = "mail",
  NOTE = "note",
  RAPPORT = "rapport",
  SOMMATION = "sommation",
  BON_COMMANDE = "bon_commande",
  DEVIS = "devis",
  DEMANDE_PARTENARIAT = "demande_partenariat",
  REJET_PARTENARIAT = "rejet_partenariat",
  ACCORD_PARTENARIAT = "accord_partenariat",
  DEMANDE_CONGE = "demande_conge",
  DEMANDE_EXPLICATION = "demande_explication",
  DEMANDE_PERMISSION = "demande_permission",
  DEMANDE_LOCATION_MATERIEL = "demande_location_materiel",
}